feat: Enable MC Dropout for model during inference (main)
This commit enables Monte Carlo (MC) Dropout for a given
PyTorch model during inference. It iterates through the
model's modules, setting all dropout layers to training
mode. This allows stochastic dropout to be applied
during inference. A flag `use_mc_dropout` is set to True
on the model to indicate that MC Dropout is active.