feat: Add MLFLOW_MODEL_PRODUCTION_NAME to docker-compose
Add the MLFLOW_MODEL_PRODUCTION_NAME environment variable to the
docker-compose.yml file. This variable is used to specify the name of the
production model in MLflow.

Remove the 'code_check_type' target from the Makefile as it is no longer
needed.

Add a 'docker_compose_build_up' target to the Makefile for convenience
when building and starting Docker containers.