feat: Use GIT_IGNORE_FILE variable in Makefile
Introduced a new variable GIT_IGNORE_FILE in the Makefile to represent the .gitignore file.
This change standardizes the reference to the gitignore file across various dvc-related targets.

The variable is now used in the following targets:
- dvc_pipeline_run
- dvc_pipeline_stage_run
- dvc_pipeline_run_force
- dvc_pipeline_stage_run_force

This makes the Makefile more maintainable and easier to update if the gitignore filename were to change in the future.