feat: Loosen network policy and allow root filesystem writes
Updates several Kubernetes network policy files to allow all ingress and egress
traffic by default. This is a temporary measure to simplify initial setup
and debugging.

Also, modifies the predictor-service deployment to set
`readOnlyRootFilesystem` to `false`. This allows the container to write
to its root filesystem, which might be necessary for certain operations
or debugging.

The CI/CD workflow is updated to ignore errors during `kubectl rollout undo`
commands. This prevents the pipeline from failing completely if a rollback
operation encounters an issue, allowing for more graceful failure handling.